+++ shotmanager/__init__.py
@@ -330,7 +330,7 @@ def register():
         print(f"\n ------ Ubisoft Shot Manager debug: {config.devDebug} ------- ")
 
     addon_prefs_inst = config.getShotManagerPrefs()
-    addon_prefs_inst.displaySMDebugPanel = True
+    addon_prefs_inst.displaySMDebugPanel = False
 
     # _props = bpy.context.scene.UAS_shot_manager_props
     # # currentLayout = props.getCurrentLayout()