+++ shotmanager/utils/utils_editors_dopesheet.py
@@ -91,21 +91,25 @@ def getPrefsUIScale():
     return bpy.context.preferences.view.ui_scale
 
 
-def getRulerHeight():
+def getRulerHeight(firstLaneIndex=0):
     """Return the height in pixels of the time ruler of a dopesheet"""
     prefs = config.getShotManagerPrefs()
-    RULER_HEIGHT = 23 * prefs.shtStack_screen_display_factor
-    # RULER_HEIGHT = 28  # on laptop
-    return RULER_HEIGHT * getPrefsUIScale()
+    RULER_HEIGHT = 23
+    # RULER_HEIGHT = 28  # on laptop at display 125%
+    rulerHeight = (
+        RULER_HEIGHT * prefs.shtStack_screen_display_factor
+    ) * getPrefsUIScale() + firstLaneIndex * getLaneHeight()
+    # rulerHeight = rulerHeight * getPrefsUIScale()
+    return rulerHeight
 
 
 def getLaneHeight():
     """Return the height of a lane in pixels"""
     prefs = config.getShotManagerPrefs()
-    LANE_HEIGHT = 18 * prefs.shtStack_screen_display_factor
-    # LANE_HEIGHT = 18.5
-    # LANE_HEIGHT = 22.5  # on laptop
-    return LANE_HEIGHT * getPrefsUIScale()
+    LANE_HEIGHT = 18
+    # LANE_HEIGHT = 22.5  # on laptop at display 125%
+    laneHeight = LANE_HEIGHT * prefs.shtStack_screen_display_factor
+    return laneHeight * getPrefsUIScale()
 
 
 # same as pixel to lane